+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add null move pruning
 * Add killer move ordering
 * Sort moves via individual move scoring
+* Sort moves incrementally
 
 ### Misc